refactor: 컴포넌트 이름변경
